authorization implemented

+++ src/api/notes/handler.js
@@ -17,8 +17,13 @@ class NotesHandler {
         try {
             this._validator.validateNotePayload(request.payload);
             const { title = 'untitled', body, tags } = request.payload;
-
-            const noteId = await this._service.addNote({ title, body, tags });
+            
+            // code ini agar kita tahu id berapa yang sedang mengakses
+            const { id: credentialId } = request.auth.credentials;
+            const noteId = await this._service.addNote({
+                title, body, tags, owner: credentialId,
+            });
+            // const noteId = await this._service.addNote({ title, body, tags, owner });
 
             const response = h.response({
                 status: 'success',
@@ -51,8 +56,10 @@ class NotesHandler {
         }
     }
 
-    async getNotesHandler() {
-        const notes = await this._service.getNotes();
+    async getNotesHandler(request) {
+        const { id: credentialId } = request.auth.credentials;
+        const notes = await this._service.getNotes(credentialId);
+        // const notes = await this._service.getNotes();
         return {
             status: 'success',
             data: {
@@ -64,6 +71,11 @@ class NotesHandler {
     async getNoteByIdHandler(request, h) {
         try {
             const { id } = request.params;
+
+            // agar mengetahui siapa yang punya notes
+            const { id: credentialId } = request.auth.credentials;
+            await this._service.verifyNoteOwner(id, credentialId);
+
             const note = await this._service.getNoteById(id);
             return {
                 status: 'success',
@@ -97,6 +109,10 @@ class NotesHandler {
             this._validator.validateNotePayload(request.payload);
             const { id } = request.params;
 
+            // agar mengetahui siapa yang punya notes
+            const { id: credentialId} = request.auth.credentials;
+            await this._service.verifyNoteOwner(id, credentialId);
+
             await this._service.editNoteById(id, request.payload);
 
             return {
@@ -127,6 +143,11 @@ class NotesHandler {
     async deleteNoteByIdHandler(request, h) {
         try {
             const { id } = request.params;
+
+            // agar mengetahui siapa yang punya notes
+            const { id: credentialId } = request.auth.credentials;
+            await this._service.verifyNoteOwner(id, credentialId);
+            
             await this._service.deleteNoteById(id);
 
             return {
